fix auto-logout form validation
