--- packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Preferences/AutoLogout/Settings/EditForm/validation.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import {
-  InvalidMessage,
-  MaximumMessage,
-  MinimumMessage
-} from './validationMessages'
-
-export const isValidAutoLogoutTime = value => {
-  if (!Number.isInteger(Number(value))) return <InvalidMessage />
-  if (value < 1) return <MinimumMessage />
-  if (value > 1440) return <MaximumMessage />
-  return undefined
-}
