+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Preferences/AutoLogout/Settings/index.js
@@ -2,26 +2,22 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 import { formValueSelector } from 'redux-form'
+import { FormattedMessage } from 'react-intl'
 
+import { Button, Text } from 'blockchain-info-components'
+import { SettingWrapper } from 'components/Setting'
+import AutoLogoutForm from './template'
 import { actions, selectors } from 'data'
-import Settings from './template.js'
 
 class SettingContainer extends React.PureComponent {
   constructor (props) {
     super(props)
     this.state = { updateToggled: false }
-    this.handleClick = this.handleClick.bind(this)
+    this.onSubmit = this.onSubmit.bind(this)
     this.handleToggle = this.handleToggle.bind(this)
   }
 
-  componentWillMount () {
-    const { logoutTime } = this.props
-    this.props.formActions.initialize('settingAutoLogoutTime', {
-      autoLogoutTime: logoutTime
-    })
-  }
-
-  handleClick () {
+  onSubmit () {
     const { autoLogoutTime } = this.props
 
     this.props.settingsActions.updateAutoLogout(
@@ -39,13 +35,27 @@ class SettingContainer extends React.PureComponent {
   render () {
     const { logoutTime } = this.props
 
-    return (
-      <Settings
-        updateToggled={this.state.updateToggled}
-        logoutTime={logoutTime}
+    return this.state.updateToggled ? (
+      <AutoLogoutForm
+        onSubmit={this.onSubmit}
         handleToggle={this.handleToggle}
-        handleClick={this.handleClick}
       />
+    ) : (
+      <SettingWrapper>
+        <Text>
+          <FormattedMessage
+            id='scenes.preferences.autologout.settings.minutes'
+            defaultMessage='{time} minutes'
+            values={{ time: logoutTime }}
+          />
+        </Text>
+        <Button nature='primary' onClick={this.handleToggle}>
+          <FormattedMessage
+            id='scenes.preferences.autologout.settings.updateform.change'
+            defaultMessage='Change'
+          />
+        </Button>
+      </SettingWrapper>
     )
   }
 }
