+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Preferences/AutoLogout/Settings/validation.js
@@ -1,23 +1,30 @@
 import React from 'react'
 import { FormattedMessage } from 'react-intl'
 
-export const InvalidMessage = () => (
+const InvalidMessage = () => (
   <FormattedMessage
     id='settings.preferences.autologout.settings.invalid'
     defaultMessage='Please set a valid time'
   />
 )
 
-export const MinimumMessage = () => (
+const MinimumMessage = () => (
   <FormattedMessage
     id='settings.preferences.autologout.settings.invalid2'
     defaultMessage='Please set a duration greater than 1 minute.'
   />
 )
 
-export const MaximumMessage = () => (
+const MaximumMessage = () => (
   <FormattedMessage
     id='settings.preferences.autologout.settings.invalid3'
     defaultMessage='Please set a duration less than 1440 minutes'
   />
 )
+
+export const isValidAutoLogoutTime = value => {
+  if (!Number.isInteger(Number(value))) return <InvalidMessage />
+  if (value < 1) return <MinimumMessage />
+  if (value > 1440) return <MaximumMessage />
+  return undefined
+}
